requirements.txt
# 资产转换依赖
pillow
pillow-avif-plugin
